public/GCP_codefiles/gcp_pdf-retrieval-function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-aiplatform
vertexai
numpy
